First prototype to establish the overall structure
This does not do much yet but it contains the whole stack that is
necessary to build this. We have

* rust->wasm compilation
* wasm/js interaction
* wasm+js+html+css bundling
* dev server with hot loading (incl. wasm)

Still missing is typescript. Tried that but it would not load the wasm
module in the js module generated by wasm-pack. Needs further
investigation first.

See the README.md file for how to build and run this project.