Implement gif reversal
This just builds on the previously established stack and provides a new
function `reverse_gif` inside the wasm module that does the heavy
lifting. This function takes a Uint8Array/slice of u8 values and returns
an array/vec of u8 values containing the reversed gif. All that is left
to do on the JavaScript side is to convert this buffer of unsigned byte
values into a data url with type `'image/gif'` and put it into an img.

Also, panic messages are now logged on `console.error()` to make the
analysis of panics a little bit easier.