Add progress reporting
Since gif encoding can take some time we want the user to know where
things are right now. So we push out an update after every frame we
encoded. This is currently only printed to the console, UI updates would
not show up anyway until the encoding process is finished since we are
currently running in the same thread as the UI is rendered in.

To fix this we have to move the gif processing into a web worker.