Process gifs inside a web worker
The gif reversal process can take a few seconds, depending on the gif
dimensions and quality. During that process no UI update is possible
because the wasm module runs in the same thread as the main UI thread.
But it would be nice to give the user some rough indicator how long it
will take to process the gif.

To do that the gif processing has to be put inside a web worker. That
way we can update the UI while the gif is processed. This would also
enable us to process several gifs in parallel by allocating multiple web
workers.

Wasm modules created with wasm-pack can't be run inside a web worker, at
least not with their wrapper script generated by wasm-bindgen because
wasm-pack only allows us to generate browser modules and node modules.
But what we need for a web worker is a wrapper script generated with
wasm-bindgen's `--no-modules` mode. So we have to use wasm-bindgen
directly after compiling the wasm module with cargo. To make this a
little bit more convenient new npm scripts are introduced. The most
important one is

    npm run build-wasm

It will call the two other scripts,

    npm run cargo
    npm run bindgen

to generate a wasm module in the ./pkg folder.