Parallel processing of gifs
This is accomplished by a pool of web workers that will process dropped
gifs in parallel. The number of web workers is dependant on the number
of available cores (`navigator.hardwareConcurrency`). That many gifs are
processed in parallel. Additional gifs are queued and processed as soon
as the first worker is finished.

The output had to be extended as well for this. Until now only a single
reverse gif was displayed. Now the output is generated for every gif and
appended to the result container div element.