Improve error handling
Until now errors inside a web worker were only logged on
`console.error()`. Now errors inside a web worker will be caught and
sent to the main thread with a message. There they can be printed to the
user to

a) show him that something went wrong and he won't get a reversed gif
b) give him the chance to report this in the github repo.